+++ src/MuMu_RichPresence/Program.cs
@@ -6,6 +6,7 @@ using Dawn.MuMu.RichPresence.MuMu;
 using DynamicData.Binding;
 using NuGet.Versioning;
 using Velopack;
+using Velopack.Sources;
 
 namespace Dawn.MuMu.RichPresence;
 
@@ -29,23 +30,21 @@ internal static class Program
     [STAThread]
     private static async Task Main(string[] args)
     {
+        Environment.CurrentDirectory = AppContext.BaseDirectory; // Startup sets it to %windir%
         Arguments = new(args)
         {
             #if DEBUG
             ExtendedLogging = true
             #endif
         };
-        Environment.CurrentDirectory = AppContext.BaseDirectory; // Startup sets it to %windir%
-        ApplicationLogs.Initialize(false);
 
         InitializeVelopack();
 
+        ApplicationLogs.Initialize();
+
         if (!Arguments.NoAutoUpdate)
-        {
-            var supportsVelopack = await AutoUpdate.Check_WithVelopack();
-            if (supportsVelopack)
-                ApplicationLogs.Initialize(true);
-        }
+            await AutoUpdate.CheckForUpdates();
+
         ApplicationLogs.ListenToEvents();
 
         SingleInstanceApplication.Ensure();
@@ -75,7 +74,6 @@ internal static class Program
     {
         var app = VelopackApp.Build();
         app.OnBeforeUninstallFastCallback(OnUninstall);
-        app.SetLogger(VelopackUpdateLogger.Create());
         app.Run();
     }